energy_mangement/auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]
python-dotenv
python-multipart
requests